ACL: Using type to distinguish user and group (#3124)
